requirements.txt
pillow
discord.py
python-dotenv
jishaku
pytz
numpy
opencv-python
pytesseract
pyocr
asyncpg